fix cluster resource
