+++ test/e2e/suite-ingress/suite-ingress-features/ingress-class.go
@@ -19,7 +19,6 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/apache/apisix-ingress-controller/pkg/id"
 	ginkgo "github.com/onsi/ginkgo/v2"
 	"github.com/stretchr/testify/assert"
 
@@ -473,81 +472,6 @@ spec:
 		assert.Contains(ginkgo.GinkgoT(), acs[0].Username, "james")
 		assert.Equal(ginkgo.GinkgoT(), map[string]interface{}{"key": "james-key"}, acs[0].Plugins["key-auth"])
 	})
-
-	ginkgo.It("ApisiClusterConfig should be ignored", func() {
-		// create ApisixConsumer resource with ingressClassName: ignore
-		acc := `
-apiVersion: apisix.apache.org/v2
-kind: ApisixClusterConfig
-metadata:
-  name: default
-spec:
-  ingressClassName: ignore
-  monitoring:
-    prometheus:
-      enable: true
-      prefer_name: true
-`
-		assert.Nil(ginkgo.GinkgoT(), s.CreateResourceFromString(acc))
-		time.Sleep(6 * time.Second)
-
-		agrs, err := s.ListApisixGlobalRules()
-		assert.Nil(ginkgo.GinkgoT(), err)
-		assert.Len(ginkgo.GinkgoT(), agrs, 1)
-		assert.Equal(ginkgo.GinkgoT(), agrs[0].ID, id.GenID("default"))
-		assert.Len(ginkgo.GinkgoT(), agrs[0].Plugins, 1)
-		_, ok := agrs[0].Plugins["prometheus"]
-		assert.Equal(ginkgo.GinkgoT(), ok, true)
-	})
-
-	ginkgo.It("ApisiClusterConfig should be handled", func() {
-		// create ApisixConsumer resource without ingressClassName
-		acc := `
-apiVersion: apisix.apache.org/v2
-kind: ApisixClusterConfig
-metadata:
-  name: default
-spec:
-  monitoring:
-    prometheus:
-      enable: true
-      prefer_name: true
-`
-		assert.Nil(ginkgo.GinkgoT(), s.CreateResourceFromString(acc))
-		time.Sleep(6 * time.Second)
-
-		agrs, err := s.ListApisixGlobalRules()
-		assert.Nil(ginkgo.GinkgoT(), err)
-		assert.Len(ginkgo.GinkgoT(), agrs, 1)
-		assert.Equal(ginkgo.GinkgoT(), agrs[0].ID, id.GenID("default"))
-		assert.Len(ginkgo.GinkgoT(), agrs[0].Plugins, 1)
-		_, ok := agrs[0].Plugins["prometheus"]
-		assert.Equal(ginkgo.GinkgoT(), ok, true)
-
-		// update ApisixConsumer resource with ingressClassName: apisix
-		acc = `
-apiVersion: apisix.apache.org/v2
-kind: ApisixClusterConfig
-metadata:
-  name: default
-spec:
-  ingressClassName: apisix
-  monitoring:
-    prometheus:
-      enable: true
-      prefer_name: true
-`
-		assert.Nil(ginkgo.GinkgoT(), s.CreateResourceFromString(acc))
-		time.Sleep(6 * time.Second)
-
-		agrs, err = s.ListApisixGlobalRules()
-		assert.Nil(ginkgo.GinkgoT(), err)
-		assert.Len(ginkgo.GinkgoT(), agrs, 1)
-		assert.Equal(ginkgo.GinkgoT(), agrs[0].ID, id.GenID("default"))
-		assert.Len(ginkgo.GinkgoT(), agrs[0].Plugins, 1)
-		_, ok = agrs[0].Plugins["prometheus"]
-		assert.Equal(ginkgo.GinkgoT(), ok, true)
-	})
 })
 
 var _ = ginkgo.Describe("suite-ingress-features: Testing CRDs with IngressClass apisix-and-all", func() {
@@ -845,78 +769,5 @@ spec:
 		assert.Len(ginkgo.GinkgoT(), acs, 1)
 		assert.Contains(ginkgo.GinkgoT(), acs[0].Username, "james")
 		assert.Equal(ginkgo.GinkgoT(), map[string]interface{}{"key": "james-password"}, acs[0].Plugins["key-auth"])
-	})
-
-	ginkgo.It("ApisiClusterConfig should be handled", func() {
-		// create ApisixConsumer resource without ingressClassName
-		acc := `
-apiVersion: apisix.apache.org/v2
-kind: ApisixClusterConfig
-metadata:
-  name: default
-spec:
-  monitoring:
-    prometheus:
-      enable: true
-      prefer_name: true
-`
-		assert.Nil(ginkgo.GinkgoT(), s.CreateResourceFromString(acc))
-		time.Sleep(6 * time.Second)
-
-		agrs, err := s.ListApisixGlobalRules()
-		assert.Nil(ginkgo.GinkgoT(), err)
-		assert.Len(ginkgo.GinkgoT(), agrs, 1)
-		assert.Equal(ginkgo.GinkgoT(), agrs[0].ID, id.GenID("default"))
-		assert.Len(ginkgo.GinkgoT(), agrs[0].Plugins, 1)
-		_, ok := agrs[0].Plugins["prometheus"]
-		assert.Equal(ginkgo.GinkgoT(), ok, true)
-
-		// update ApisixConsumer resource with ingressClassName: apisix
-		acc = `
-apiVersion: apisix.apache.org/v2
-kind: ApisixClusterConfig
-metadata:
-  name: default
-spec:
-  ingressClassName: apisix
-  monitoring:
-    prometheus:
-      enable: true
-      prefer_name: true
-`
-		assert.Nil(ginkgo.GinkgoT(), s.CreateResourceFromString(acc))
-		time.Sleep(6 * time.Second)
-
-		agrs, err = s.ListApisixGlobalRules()
-		assert.Nil(ginkgo.GinkgoT(), err)
-		assert.Len(ginkgo.GinkgoT(), agrs, 1)
-		assert.Equal(ginkgo.GinkgoT(), agrs[0].ID, id.GenID("default"))
-		assert.Len(ginkgo.GinkgoT(), agrs[0].Plugins, 1)
-		_, ok = agrs[0].Plugins["prometheus"]
-		assert.Equal(ginkgo.GinkgoT(), ok, true)
-
-		// update ApisixConsumer resource with ingressClassName: watch
-		acc = `
-apiVersion: apisix.apache.org/v2
-kind: ApisixClusterConfig
-metadata:
-  name: default
-spec:
-  ingressClassName: watch
-  monitoring:
-    prometheus:
-      enable: true
-      prefer_name: true
-`
-		assert.Nil(ginkgo.GinkgoT(), s.CreateResourceFromString(acc))
-		time.Sleep(6 * time.Second)
-
-		agrs, err = s.ListApisixGlobalRules()
-		assert.Nil(ginkgo.GinkgoT(), err)
-		assert.Len(ginkgo.GinkgoT(), agrs, 1)
-		assert.Equal(ginkgo.GinkgoT(), agrs[0].ID, id.GenID("default"))
-		assert.Len(ginkgo.GinkgoT(), agrs[0].Plugins, 1)
-		_, ok = agrs[0].Plugins["prometheus"]
-		assert.Equal(ginkgo.GinkgoT(), ok, true)
 	})
 })
